v1.0.0 - Switch to TypeScript
