--- .travis.yml
@@ -1,7 +1,0 @@
-language: node_js
-node_js:
-  - 'stable'
-  - 'lts/*'
-  - '12'
-after_success:
-  "npm run coverage"
