--- index.js
@@ -1,13 +1,0 @@
-/*! Copyright Kuu Miyazaki. SPDX-License-Identifier: MIT */
-const {getOptions, setOptions} = require('./utils');
-const parse = require('./parse');
-const stringify = require('./stringify');
-const types = require('./types');
-
-module.exports = {
-  parse,
-  stringify,
-  types,
-  getOptions,
-  setOptions
-};
